+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.5.3 - 2023-05-18
+### Fixed 
+- Added fallback for block elements without a created structure to fix content provider
+
+## 3.5.2 - 2023-05-17
+### Changed 
+- Updated error message for failed jobs after retries
+
+### Fixed
+- Resolved duplication issue with Neo and SuperTable fields
+
 ## 3.5.1 - 2023-04-04
 ### Fixed
 - Query for fetching translations by status and id